69. reduce peak RSS ~50% with uint16 merged, rfftn, and integer roll fast path (Phase C perf)
- fov.py: _make_ref_3d uses dtype=np.uint16 (was int64, saves 3.2 GB)
- phase_correlate: rfftn/irfftn replaces fftn/ifftn (~50% less FFT memory)
- apply_shift: np.roll + zero-fill for integer shifts (no FFT round-trip)
- Benchmark: tissue_2D 19.9→11.0 GB, cell_culture_3D 5.5→2.8 GB, ~2x faster

+++ src/python/starfinder/dataset/fov.py
@@ -224,7 +224,8 @@ class FOV:
         """Create 3D reference/moving image for registration."""
         img = self.images[round_name]
         if mode == "merged":
-            return np.sum(img, axis=-1)
+            # uint16 is sufficient: max sum of 4 uint8 channels = 1020
+            return np.sum(img, axis=-1, dtype=np.uint16)
         else:
             return img[:, :, :, channel]
